+++ lorm/tests/resources/migrations/mysql/12_drafts_table.sql
@@ -1,0 +1,5 @@
+CREATE TABLE IF NOT EXISTS drafts (
+    id      BINARY(16) PRIMARY KEY NOT NULL,
+    title   TEXT       NOT NULL,
+    user_id BINARY(16)
+);